% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v22.R
\name{v22}
\alias{v22}
\title{Calculates v22 predict survival.}
\usage{
v22(age = 66, size = 12)
}
\arguments{
\item{age}{Age at diagnosis}

\item{size}{Invasive tumour size (mm)}
}
\value{
A tibble
}
\description{
Calculates v22 predict survival.
}
\keyword{breast}
\keyword{cancer}
\keyword{survival}
